not quite finished

+++ app/controllers/sessions_controller.rb
@@ -3,35 +3,13 @@ class SessionsController < ApplicationController
   before_filter :signed_login_required, only: [:abort_signed_session, :redirect_to_asigned]
   before_filter :token_login_required, only: [:abort_token_session]
 
-  def redirect_to_asigned
-    # raise session.to_yaml   
-
-    case session[:user_signed][:role]
-    when "student"
-      redirect_to homepage_student_path
-    when "profesor"
-      # require 'net/http'
-      # require 'json'
-      # user_url = URI.parse("http://fmi-autentificare.herokuapp.com/users/#{session[:user_signed][:uid].to_i}.json")
-      # response = Net::HTTP.get_response(user_url)
-      # user = JSON.load response.body
-      
-      redirect_to homepage_profesor_path
-    when "administrator"
-      redirect_to homepage_admin_path
-    else
-      flash[:error] = "Rol nesuportat de aplicatie : #{session[:user_signed][:role]}"
-      redirect_to sign_out_path
-      #abort_signed_session
-    end
-  # rescue
-  #   flash[:error] = 'Ceva neobisnuit s-a petrecut..'
-  end
-
+  # pagina de login
   def new
     redirect_to verificare_path if session[:user_token]
+    redirect_to homepage_path if session[:user_signed]
   end
 
+  # logare prin token pentru evaluarea cursurilor
   def create
     if params[:sessions][:token].blank?
       flash[:error] = "Nu te poti loga cu un camp gol!"
@@ -40,6 +18,13 @@ class SessionsController < ApplicationController
 
     i_user = IncognitoUser.find_by_token params[:sessions][:token].strip
     if i_user
+      # verifica daca e perioada de evaluare
+      data_eval =  i_user.grupa.data_evaluare.data
+      unless Date.today.between? data_eval, data_eval + GossipLogin::DURATA_PERIOADA_EVALUARE - 1
+        flash[:error] = "Momentan nu puteti evalua nici un curs"
+        render 'new' and return
+      end
+      
       # mentinem utilizatorul
       session[:user_token] = {token: i_user.token, grupa: i_user.grupa_nume}
 
@@ -59,7 +44,7 @@ class SessionsController < ApplicationController
       end
 
     else
-      flash[:error] = "Acel token e invalid.... si eu ma plictisesc"
+      flash[:error] = "Acel token e invalid...."
       render 'new'
     end
 
@@ -67,44 +52,78 @@ class SessionsController < ApplicationController
 
 
   def abort_token_session
-
-    # mai bine lasam in baza, nu incurca pe nimeni
-    sesiune = SesiuneActiva.find_by_incognito_user_token(session[:user_token][:token])
-    sesiune.destroy if sesiune
-
-    session[:user_token] = nil
-
-    # deci nu va plac mesajele cu stil
-    #flash[:notice] = "Ai iesit cu ochii-n soare"
-
+    clear_session
     redirect_to root_path
   end
 
   def abort_signed_session
-    session[:user_signed] = nil
+    clear_session
     redirect_to root_path
   end
 
+  # logare prin FMI-Connect
   def create_signed
     # raise env["omniauth.auth"].to_yaml
-    
     session[:user_signed] = { uid: env["omniauth.auth"]["uid"],
                               token: env["omniauth.auth"]['credentials']['token'],
                               first_name: env["omniauth.auth"]["extra"]["first_name"],
                               last_name: env["omniauth.auth"]["extra"]["last_name"],
-                              role: case 
-                                    when env["omniauth.auth"]["extra"]["student"] then 'student'
-                                    when env["omniauth.auth"]["extra"]["teacher"] then 'teacher'
-                                    # when env["omniauth.auth"]["extra"]["management"] then 'sef_def'
-                                    when env["omniauth.auth"]["extra"]["admin"] then 'admin'
-                                    else nil 
-                                    end}
-    
+                              role: role_extract(env["omniauth.auth"]["extra"])}
     redirect_to_asigned
   end
 
-  def failure
-    flash[:notice] = params[:message]
+  # titlu bine ales
+  def redirect_to_asigned
+    case session[:user_signed][:role]
+    when "student" then
+      redirect_to homepage_student_path
+    when "profesor" then
+      redirect_to homepage_profesor_path
+    when "admin" then
+      redirect_to homepage_admin_path
+    else
+      flash[:error] = "Rol nesuportat de aplicatie"
+      redirect_to sign_out_path
+    end
+  end
+
+
+  # urmeaza metode private
+  private
+
+  # pentru ca 
+  def role_extract(extra)
+    case 
+    when extra["student"] && (extra["student"] == true  || extra["student"] == 'true') then 'student'
+    when extra["teacher"] && (extra["teacher"] == true  || extra["teacher"] == 'true') then 
+      sef_dep = false
+      begin
+        require 'net/http'
+        require 'json'
+        user_url = URI.parse("http://fmi-autentificare.herokuapp.com/users/#{session[:user_signed][:uid].to_i}.json?oauth_token=#{session[:user_signed][:token]}")
+        response = Net::HTTP.get_response(user_url)
+        user = JSON.load response.body
+        sef_dep = true if user['user']['teacher']['is_responsable'] == 1
+      rescue
+        flash[:error] = "Nu am putut determina daca sunteti sef departament"
+      ensure
+        sesiune[:user_signed][:sef] = sef_dep
+        'teacher'
+      end
+    when extra["admin"] && (extra["admin"] == true  || extra["admin"] == 'true') then 'admin'
+    when extra["administrator"] && (extra["administrator"] == true  || extra["administrator"] == 'true') then 'admin'
+    else nil
+    end
+  end
+
+  def clear_session
+    if session[:user_token]
+      sesiune = SesiuneActiva.find_by_incognito_user_token(session[:user_token][:token])
+      sesiune.destroy if sesiune
+      session[:user_token] = nil  
+    end
+
+    session[:user_signed] = nil
   end
 
 end
